// src/app/shared/modules/tables/model/tables.config.ts
export enum TableCellTypes {
  link = 'link',
  date = 'date',
  dateMonthYear = 'dateMonthYear',
  counter = 'counter',
  email = 'email',
  status = 'status',
  currency = 'currency',
  additionAndDeduction = 'addition_deduction',
  secret = 'secret',
  extraInfo = 'extraInfo',
  eventEmitter = 'eventEmitter',
  list = 'list',
  listOfNames = 'listOfNames',
  listOfErrors = 'listOfErrors',
  toggle = 'toggle',
  linkWithExtraInfo = 'linkWithExtraInfo',
  checkbox = 'checkbox',

}

export enum TableCellAligns {
  start = 'start',
  end = 'end',
  center = 'center',
}